+++ instrumentation/github.com/gorilla/mux/otelmux/mux.go
@@ -19,6 +19,7 @@ import (
 	"net/http"
 	"sync"
 
+	"github.com/felixge/httpsnoop"
 	"github.com/gorilla/mux"
 
 	otelcontrib "go.opentelemetry.io/contrib"
@@ -74,27 +75,6 @@ type recordingResponseWriter struct {
 	status  int
 }
 
-func (w *recordingResponseWriter) Header() http.Header {
-	return w.writer.Header()
-}
-
-func (w *recordingResponseWriter) Write(slice []byte) (int, error) {
-	w.writeHeader(http.StatusOK)
-	return w.writer.Write(slice)
-}
-
-func (w *recordingResponseWriter) WriteHeader(statusCode int) {
-	w.writeHeader(statusCode)
-	w.writer.WriteHeader(statusCode)
-}
-
-func (w *recordingResponseWriter) writeHeader(statusCode int) {
-	if !w.written {
-		w.written = true
-		w.status = statusCode
-	}
-}
-
 var rrwPool = &sync.Pool{
 	New: func() interface{} {
 		return &recordingResponseWriter{}
@@ -105,7 +85,26 @@ func getRRW(writer http.ResponseWriter) *recordingResponseWriter {
 	rrw := rrwPool.Get().(*recordingResponseWriter)
 	rrw.written = false
 	rrw.status = 0
-	rrw.writer = writer
+	rrw.writer = httpsnoop.Wrap(writer, httpsnoop.Hooks{
+		Write: func(next httpsnoop.WriteFunc) httpsnoop.WriteFunc {
+			return func(b []byte) (int, error) {
+				if !rrw.written {
+					rrw.written = true
+					rrw.status = http.StatusOK
+				}
+				return next(b)
+			}
+		},
+		WriteHeader: func(next httpsnoop.WriteHeaderFunc) httpsnoop.WriteHeaderFunc {
+			return func(statusCode int) {
+				if !rrw.written {
+					rrw.written = true
+					rrw.status = statusCode
+				}
+				next(statusCode)
+			}
+		},
+	})
 	return rrw
 }
 
@@ -145,7 +144,7 @@ func (tw traceware) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	r2 := r.WithContext(ctx)
 	rrw := getRRW(w)
 	defer putRRW(rrw)
-	tw.handler.ServeHTTP(rrw, r2)
+	tw.handler.ServeHTTP(rrw.writer, r2)
 	attrs := semconv.HTTPAttributesFromHTTPStatusCode(rrw.status)
 	spanStatus, spanMessage := semconv.SpanStatusFromHTTPStatusCode(rrw.status)
 	span.SetAttributes(attrs...)